Update main.py
Fixed a bug where damage caused in battle lasted outside of battle and added the option to sell 10 HP for ￥5,000. Until adding more items to the store, this may be a useless option, but it definitely adds to the eeriness in that you can sell your life away for money.
Also removed unnecessary 'from time import sleep' line.